Completely removed templates from image set code.
Added function to check filenames to see which probably belong in the same image set. Whilst the ImageSet class no longer contains templates, the template_regex stuff is used to check similarity here. Files must be of a form {PREFIX}_{SOME_NUMBERS} otherwise they are assigned to their own ImageSet. We could probably make this matching a lot smarter if we needed to.

Also added check to see whether image set is a sweep by sequential indices.

Added a factory function that, given a list of filenames, returns a list of imagesets/sweeps.