Don't rely on unspecified behaviour of std::polar implementation when passing negative values of rho.
In the case of passing a negative value of rho to the libc++ implementation of std::polar, this function
returns NAN. This appears to be in keeping with the definition of the std::polar function which
specifies [1,2]:

Returns: The complex value corresponding to a complex number whose magnitude is rho and whose
phase angle is theta.

As of Mac OS X 10.9, libc++ is now the default version of the standard library with XCode.

[1] http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2012/n3337.pdf
[2] http://lists.cs.uiuc.edu/pipermail/cfe-dev/2013-November/033591.html