1) FormatCBF no longer relies on a binary section.  Allows CBFs that are only headers to be read.
2) Finish wiring up optical2cbfheader.  Reads an optical 
metrology flat file and writes a cbf header describing the
detector geometry.  Sets distance to zero.

3) cxi.pickle2cbf accepts cbf files as input metrology, which
copies the geometry from the input and applies it to the out
put.  Currently glitched: it uses the distance from the input
cbf metrology when it should retain the distance from the
pickle image.