First pass at implmenting split sigma test as defined by Peter Zwart.
ssr = sum( (Iha-Ihb)^2 ) / sum( sigma_ah^2 + sigma_bh^2)

where Iah and Ibh are merged intensities for a given hkl from two halves of
a dataset (a and b). Likewise for sigma_ah and sigma_bh.

ssr (split sigma ratio) should approximately equal 1 if the errors are correctly
estimated.

This commit merely reports ssr as SpSig in the final table.  The next commit
will apply the ratios to the sigmas according to the binning requested during
merging.