Make CIF crystal structure builder resilient to missing atom type
Some CIF files feature atom_site_label but not atom_site_type_symbol.
It is important to support this case as it concerns over 15,000
structures in the 9th directory of the COD which came from the
American Mineralogist Crystal Structure Database.

The fix is simple: just let xray.scatterer constructor guess the
atome type from the atom label.