XFEL DB
Schema updates:
Add extra_phil_str to to rungroup table.  Not hooked up yet, but idea is to allow user to add run-group specific phil bits.  Will be needed for upcoming experiment.
Add rungroup_id to event table, see below.
Rename detector.sifoil to detector.attenuation. Not used at present.

Fixed subtle bug where images in active rungroups were being included when tallying number of indexed images in a trial.  To fix, had to add rungroup_id to event table.  To make this a non-breaking change, a few bits are checked in xfel/ui/db before reading the new table to see if the column exists.  Thus, the tallying bug will be fixed only if the user regenerates their tables, but if they don't they should still be able to use the gui.

To support this fix, several changes are made in command_line to pass rungroup on to the db logger to be included in each event entry.