IOTA v1.0.015 - Monitor Mode
Features:

1. Monitor mode: when toggled, IOTA will wait for new images to appear in
the indicated input folder (or subfolders of that) and then process them.
2. New "Preferences" (GUI-only) allow to pre-select monitor mode and to
set a time-out timer.
3. Also added "Monitor" and "Timeout" buttons in the processing window,
allowing user to toggle those options on and off. A separate dialog pops up
when "Timeout" is toggled, presenting some timeout options.

(This feature has not yet been tested with live data collection.)