Metal Coordination Library - SF4
Add the beginnings of a metal coordination library using SF4 as the
first entry. Includes adjustment of input coordinates (really atom
names mostly) to match the ideal geometry and restraints; specific
coordination of the cluster to CYS SG with more general to come;
blocking the old automatic linking code from coordinating SF4; and
adding a matching set of restraints to GeoStd.