Prediction multiplicity calculation in merging.
Enable calculation of prediction multiplicity, a measure of multiplicity
based on all predictions to the edge of the detector. When these
predictions are not found in the integration pickles, but an example
image is supplied (for pixel size and active areas), the predictions are
generated and their miller indices are tracked throughout merging.
Prediction multiplicity calculation is enabled with the phil parameter
predictions_to_edge.apply = True.