Rejigger FormatCBFCspadInMemory.
FormatCBFCspadInMemory is a Format class that doesn't need to be instantiated with a file name string, but instead is loaded using an instantiated cbf_handle.  This commits renames FormatCBFCspadInMemory to FormatCBFFullStillInMemory to make it generic, then re-adds FormatCBFCspadInMemory as a mixin class, mixing FormatCBFFullStillInMemory and FormatCBFCspad.

Intent is to use FormatCBFFullStillInMemory to work with Rayonix data from XTC streams in a subsequent commit.