NeXus changes
Change check_shape to handle a list of possible shapes.  Allows data arrays with 2 or more dimensions in the case of a hyperslab of data. For example, the CSPAD is stored in arrays with dimension Nx8x185x388. Adjust where check_shape is called to handle this.

Handle angstrom and angstroms in convert_units