Revert some of the import changes
This is more complex a dependency web than it appears, because several
instances were (inadvertantly?) hacking around python placing the argv-specified
script onto the PYTHONPATH when tests were placed in the scitbx module root,
which means both relative *and* absolute importing was broken.

Also an mmtbx module relied on relative importing - it appears to just get a
module from the current location without a package-level specifier. This
has been changed to explicit-relative syntax.