Only write changes to files if syntax is still correct and the resulting syntax tree does not change. Properly don't visit .git, python cache, and jinja directories.
A syntax check is relevant when adding the 'print_function' import. It prevents breaking code, eg.
  from __future__ import print_function
  print "string"

The syntax check alone still does not guarantee working code. For example
  from __future__ import print_function
  with open('x', 'w') as f:
    print >> f, ""
is valid syntax and successfully parses as tuple(function right-shifted by a file handler, "").
Which of course is unsemantic nonsense. These cases can be caught by comparing the syntax trees
before and after the change.