Add flag to libtbx.configure to enable Python3 migration warnings.
This patch is part of the ongoing project to migrate to Python3.
Python2 has a command line option to show warnings when code is run
that may not migrate cleanly to Python3. These warnings can now be
enabled (across all dispatchers) by running

   libtbx.configure --python3warn=warn libtbx

and disabled again with

   libtbx.configure --python3warn=none libtbx

The warnings are disabled by default.

You can also enforce python3 compatibility by stopping execution
when a warning is encountered. This may be useful for running
tests in the future. (Currently you will not get to the point
where you can run tests.)
This mode can be enabled with

   libtbx.configure --python3warn=fail libtbx