Special case for cspad.cbf_metrology: flat_refinement_with_distance
If True, and if using flat refinement, then use constraints to allow disance to refine at levels 1 and higher. For example, if refining inner 8 sensors of the CSPAD, allow a single parameter to control the distance value using a constraint.

Because of this, unrefined panels will be 'left behind'.  After refinement, reload the refined model and sync the distance change to the unrefined panels.