Change initial values for sdfac_refine
In the normal probability analysis, remove all zeros first. These are reflections with only one value so the normalized error will be zero.
Also, use sqrt(sdadd) as initial value instead of 0 for sdb.

(the intial values for the sdfac parameters appear to not affect the simplex minimizer, but they do affect the upcoming lbfgs one)