Add an C++17-style allocator model to array_family
A custom allocation model allows control of where and how the
array_family instances allocate memory. The C++17 model was chosen
because it replaces the complex, unfit-for-purpose C++03 model with
something much cleaner and easier to understand. See §23.12 in the C++17
standard for details on the actual API.

Default behaviour should be completely unchanged, but all of the
allocating array_family models now have an additional "allocator"
argument to their constructors. By default, they use the current
'global' allocator, which defaults to something that just implements
standard C++ new/delete, but can be changed to something custom.

Due to C++03 limitations the API is not complete and several advanced
behaviours have been left out - some would be possible with excessive
template metaprogramming, but weren't deemed essential to get the
fundamental behaviour. In particular, the following are not supported:
- Automatic forwarding of scoped allocators via contruct
- Alignment specification. Some parts of the API allow specification of
  alignement separately from the C++11 alignas type specifiers, but these
  are not currently implemented. Alignment == max_align is asserted for,
  but only in debug builds.
- Specialized allocators (synchronized_pool_resource,
  unsynchronized_pool_resource and monotonic_buffer_resource) are not
  implemented.

In addition, because of the model of array_family::shared - e.g. normal
C++ copying semantics have been bypassed in order to pass a shared
reference around, there is an additional restriction: Copy-assignment
only works between objects with identical allocators; otherwise this
reference share would have to become a potentially expensive (and
misleading) copy. Most, if not all cases of this should be avoidable
with careful management of object construction. The same argument
applies to ‘swap’.

References:
- Pablo Halpern; "Allocators: The Good Parts". CPPCON17. https://youtu.be/v3dz-AKOVL8
  Good summary of the arguments and reasoning behind the changes, by the actual author
- Original proposal: https://isocpp.org/files/papers/N3916.pdf
- Accepted TS: http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2015/n4480.html