Revert "Python3 syntaxification, #80"
This reverts commit c3d62324b7ff1fe191fd52a7bec105b6b9ae95fa.

Turns out libtbx runs other Python code not by importing it (which would be the normal way), but by simply executing the file in the local context.
This assumes the file to be of the same syntax as the current file, therefore leaking from __future__ import print_function to said file.

In other words:
 1. all libtbx_refresh.py need to be made Py3 compatible before proceeding.
 2. execfile() doesn't exist in Python3 anyway so this needs to change.
 3. this will affect all cctbx modules.
 4. we should really not roll our own package management system.