Alter copy-construction of dxtbx::model::Detector
Previously, the Panel superclass part of the Node class was created with
default initialisation, and then rewritten with copy-assignment. This
changes that to avoid the additional assignment - the Node is constructed
with the desired Panel information in the first place. This is required 
to avoid potential problems with an allocator implementation project.

Also moved an internal copy routine to private, and clarified it’s documentation.