Generate dispatchers for all console_scripts entry points
console_scripts entry points is the standard mechanism to
create executable commands. Here we generate libtbx dispatchers for
all active registered console_scripts that exist in the base/bin
directory. This only works on cctbx installation with a base python.
On base-less installations (and on Windows) nothing will be generated.

This addresses #145 for python commands, but not 3rd party executables.