Python3 conversion, #80
cStringIO is gone. I suspect the extra pickling bits are no longer
relevant, but keep them anyway and replicate the cStringIO readonly
behaviour. (Function is only used in 2 places anyway and should be
scrapped)

Although six has been in the base build for some time now, allow
for systems that do not have six yet (as this is the first time
we use it in this repository). It will be required sooner or later,
so make sure you have it in base-less builds.