Enable striping across multiple rungroups.
Retain default behavior of striping or chunking within a rungroup, but
add a phil parameter to override this in the case that several rungroups
are known to have the same sample-to-detector distance, and allow
striping across the rungroup barriers. This behavior is not affected
by selection of rungroups to include (or lack thereof).