tst_nanoBragg_pilatus.py will create a Pilatus-like image, complete with bad pixels, window panes, thickness effects, module-to-module misalignment and 0.1 deg of phi rotation.  It takes one optional argument, which is the image number in a sequence of 3600.  I've processed this simulated data set with xia2 and XDS and found the anomalous signal has the right sign.
This required some edits and additions to the simTBX code as well as FormatCBFMini.as_file() that I think fixed a beam center consistency bug.  Now if you dxtbx.load() and FormatCBFMiniPilatus.as_file() you get the same beam center in both headers.  I also enhanced the header to make it look more like the ones in DIALS regression.  A new "SIM.F000" value is now available for marking the "direct beam" on the image itself.