Simplify python pip builder
Only do download->install when there is a step to be done between download and install.
Otherwise go straight for install. Download->install may cause prerequisites to be downloaded
that are then not required for installation.
There is only one package that as a callback_before_install method, but that one doesn't use
pip yet, so it is unclear whether download->install is required at all.