Not OK: remove "print OK" from all dxtbx tests.
People can, and do, assume that 'print OK' means the test succeeded, and then
logically deduce that when 'OK' is not printed that must surely mean the test
failed. So they start writing tests that are supposed to fail if 'OK' is not
printed. However the testing frameworks (both pytest and libtbx) do not care
whether 'OK' is printed, so the test will always suceed.

This is one reason why we may get fundamentally broken tests.