Replace deprecated boost::python numpy interface
Boost 1.63 deprecated boost::python::numeric and added ::numpy. In order
to update to 1.65 (#172), which drops the numeric API, we needed to move
to the new ::numpy.

Because the numpy bridge code is mainly written in plain numpy C-API,
the opportunity has been taken to update the API syntax (to remove
warnings about using the old API).

The old macro-based declaration scheme has been removed in favor of C++
templates, which means the numpy conversion routines can be specified
more centrally for flex arrays, though for some reason this appears to
fail for std::size (which is why that is overridden specifically in
flex_size_t.cpp)

Fixes #70.