Python 3 boost.python injector change
The boost.python.injector depends on the behaviour of __metaclass__.
This has changed in python 3, so boost injections using the class
constructor will no longer happen. There are no error messages or
anything - the code is simply never executed.

To work around this a new method inject() was added to boost.python.
It takes an extension class as the first parameter and any number of
further classes, the methods of which are then injected into the
extension class, as before.