Generate libtbx.* wrappers from correct directory
Python console scripts don't necessarily land in base/bin, specifically
on MacOS this is not the case. Use setuptools etc. to find the actual
Python console script location and use that directory instead. Fixes #177

Add a function to regenerate all Python console scripts which can be
called from an installer to relocate paths.