Enable compilation of boost::numpy on Windows
This is a necessary prerequisite to #70.

Reason this failed was that BOOST_PYTHON_SOURCE was defined for the
boost::numpy components. This led to an attempt to export boost::python
in the boost::numpy library, which then failed during linking.

On Windows boost::numpy requires boost 1.64.
This is due to a bug in 1.63 which means the boost_numpy.lib is not
created. https://github.com/boostorg/python/issues/98