Refactor FormatXTC and derived classes to remove class variables _src and _img
Theses class variables are scoped globals which cause MPI broadcasting of datablock objects to fail.  For XTC streams, having only one rank call populate_events is important to limit file reads, so MPI programs should read the datablock once, then broadcast it to other ranks. Can't rely on _src and _img to be populated in that case in the other ranks.