Add test for split datablocks. Test is marked as expected to fail.
The test reads an hdf5 file as a datablock, splits it so the datablock contains one image, then serializes the datablock.  Deserializing the datablock with check_format=True works fine.  The single file indices array, which tracks which images are in the datablock, is correctly read from the json.  However, check_format=False does not correctly deserialize the array.  This affects all cases of check_format=False, noteably in dials.split_experiments and dials.combine_experiments.