XFEL GUI: plot resolution on the RunStats tab
Switch the I/sig(I) plot over to a resolution plot. The third subplot
used to plot I/sig(I) but this is hardly ever informative in practice.
Instead plot the resolution where I/sig(I) is at least the cutoff value,
default of 1. The "high quality" rate is now the rate at which this
resolution is at least the d_min, which should be equivalent to the old
behavior.

TODO: make the resolution y axis distributed so that high resolution
stands out better.