libtbx.find_clutter: can now specify files that should be ignored using .clutterignore
When checking for clutter, the program will first look for a file named .clutterignore in the same path.  If present, and if a line in .clutterignore matches the file name, the file will not be checked for clutter.

Useful for regression folders with files we do not wish to change.