Temporarily revert "Add an C++17-style allocator model to array_family"
Merging this has exposed problems with the windows implementation. Since
a Phenix release is in process and windows is a supported platform, it
seems more prudent to remove this from the master branch and take time
studying the problem more thoroughly rather than trying to rush or hack
something in for a release.

This reverts commit 43eeebc714732354375188a065e44abebc0f298e.
This reverts commit 50298f07d4b10335cb72bfd5b9b025dd53ca983a.