Merge image reading tests with dials_regression
There was a test in test_dxtbx.py that read some subset of
dials_regression/image_examples, and a test in dials_regression that
read most of the images that it could find. This replaces the former
with a pytest version of the latter. The original
dials_regression/test.py test will remain for now. Also includes the
dials_regression/tst_all_formats.py tests.

Also, removed the (skipped) compression tests; these are probably better
tested with explicit compressed image files and/or a unit test of the
functionality - compressed image reading is not handled at the
per-Format level.