Disable failing test
Test was recently added in 571769f1435f294b5dc71a082a076ac1887ccb38 as a
known broken test.
Right now we need a stable build to branch DIALS 1.11.

Unfortunately libtbx testing does not support "This test is failing, but
that is OK, run it anyway", so I am simply commenting out the test.