Increase specificity of FormatBrukerFixedChi
Bruker Photon II images should be read by FormatBrukerPhotonII.
Unfortunately, FormatBrukerFixedChi understood them too if they
have 'FixedChiStage' in the header. This commit settles that
dilemma, but in the ugly way of requiring that FormatBrukerFixedChi
explicitly rejects Photon II images.

A better fix should be possible as part of the resolution of #65.