Add flag slow_tests=False to libtbx.run_tests_parallel
This flag will cause the program to look for additonal tests in a list named tst_list_slow, and if it finds any, to add them to the list of tests to be run.

Additional notes have been added here:
https://github.com/cctbx/cctbx_project/wiki/cctbx-Developer-Guidance

Thusly:
"Slow tests. Slow tests may be marked as such by adding them to a list named tst_list_slow in the run_tests.py file. These tests will not be ran by default, but can be run by adding the flag slow_tests=True to libtbx.run_tests_parallel. Slow tests should only be added with good reason, such as if 1) they are designed to test large problems that will always run slowly, such as constructing and inverting large matrices or 2) they are designed to reproduce results from a publication."

xfel_regression will be adding some slow tests soon that we don't want to be ran on the command line but we do want to be ran by buildbot. Hence this new feature.