Installer: Force use of provided script
When generating an installer we automatically generate two
   ./xxxxx_env.sh and
   ./xxxxx_env.csh
scripts in the top level directory of the installer archive.
If the installer generation is run with --installer_script
then the given script ends up as
   ./install
in the same place.

The user now has two options: They can run the install script
or they can source one of the shell scripts.

While the result *may* be the same, we have now identified one
case where this led to a runtime bug which was very difficult
to track down.

This patch changes the default behaviour to not generate the
additional shell scripts if an installer script is provided.