List prime as an optional dependency of xfel.
In #247, it was pointed out that xfel/clustering/singleframe.py imports from prime.api, and it was proposed to move prime.api to xfel to break the dependency. However the xfel gui also uses prime, so it is not practical to break the circular dependency by moving code.  Since prime has a hard requirement on xfel, resolve the circular dependency by making prime an optional dependency of xfel.

Closes #247.