Bugfix for ImageSetLazy
If operator[] or get_detector() hasn't been called on an image in a Lazy image set, then get_corrected_data (in imageset.h) will fail as the detector isn't loaded yet.  Override get_corrected_data and get_gain to explictly load the dxtbx models before calling the C++ code.

Probably not a bug we've noticed as either we read the data with operator[] or get_detector() is called before get_corrected_data().

Thanks to @dermen for noticing the bug.