Add optimise_for_sparse option to lstbx.normal_equations.add_equation
This commit adds an option to use choose the method for calculating
the A^T W A matrix during add_equation, to enable switching depending
on the shape and sparsity of the problem. The default is set the
new behaviour after pull request #295. Setting
optimise_for_sparse=False enables one to recover the previous
behaviour.