Fix issue with clang compiler
Parallel branch execution in recent versions of clang causes the branch in the test 'if (x > 3)' to be executed even if x <= 3. If w <= 0 this causes a log(0) floating point exception which is caught by the boost FPE traps.

Adding an assert matches other programming patterns in cctbx and ensures that even in parallel branch execution, log(0) is not called.

Fixes #371

(thanks @rjgildea)