XFELGUI: use Cheetah instead of custom facility for SACLA
I've preserved the dbpy run finder as a comment in case we want to use it again. Otherwise, this commits reverts the SACLA facility as an XFEL GUI option, since processing at SACLA should be done in Standalone mode using Cheetah to dump the h5 files. Since we also have PBS support, if we are only using Cheetah we have everything we need.