The second point is the first non-collinear point from the list
if any point is collinear then it can be dropped from the list as it
would also be collinear with the third point, so won't be considered in
the future.
The second point itself also can't be identical to the third point, as
the triplet would then always be coplanar, so again can use .pop() here.