Drop the sort comparison helper function
instead sort by a key function using a tuple. The first element is the
vector length (distance), and we sort by those from low to high. If the
distance is identical then the second element is used, which is the sum
of the vector elements, but here we sort from high to low, so invert the
value.

(NB: The comment in the helper function was misleading. This was not for
purely aesthetic reasons at all. If the sorting is not done this
specific way the code will either crash or produce different output)