This function is always the identity.
We know that no item in the list is collinear to the item we want to add
- otherwise we would have run into 'continue' already.

We know that the items we are looking at are sorted with increasing
distance.

Therefore if the vector we want to add had all components as a multiple
of 2 then we would have encountered the vector with the reduced components
before.

Which would then be collinear to the vector we are looking at, so we
would not be looking at the vector in the first place.

Proof by contradiction: the components of the vector can not all be
multiples of 2.

Therefore _divide() always returns its original function parameter.

QED.