Optionally avoid auto_anomalous check
Add an anomalous=None parameter to iotbx.mtz.object.as_miller_arrays().
If this is set to True or False, then the call to milller.set.auto_anomalous()
in iotbx.mtz.column_group() is avoided. This call can potentially be expensive
with particularly large reflection files with millions of observations.
In iotbx.merging_statistics pass the value of the anomalous parameter to
iotbx.mtz.object.as_miller_arrays().

Fixes #321