get_top_solutions bugfix for HemisphereSampler
HemisphereSampler implements the Steller 1997 algorithm for choosing candidate basis vectors. Its method get_top_solutions is supposed to sort the putative basis vectors by their scores (kval), then pick the top 30 that are unique directions.  However, the current implementation keeps all non-unique vectors as it searches for unique ones, due to an indentation error in the python code. In some cases, this can lead to it returning thousands of candidates.

The algorithm iterates through each candidate vector and tests if it is within 5 degrees of an already tested vector.  The bugfix here changes the indentation of hemisphere_solutions_sort.append(test_item) to only occur if the tested direction is actually unique.

This change affects existing indexing results and tests, so corresponding changes to tests in extant libraries (e.g. DIALS) need to be made.

(this is a long-standing bug)

Co-authored-by: Nicholas Sauter <nksauter@lbl.gov>