Catch comparisons of a NoneType with a float
This appears to happen if the number of bins is small (perhaps if there
is only a single bin?), which only happens when there are very few
reflections.  The intended logic here is fairly opaque to me, but this
fix seems minimally invasive.

Also removed some redundant parentheses and unused variables.