Fix check that an array is not already merged
If we create an instance of dataset_statistics with anomalous=False and
pass an intensity array i_obs that is not a unique set under symmetry,
but whose corresponding anomalous array is unique under symmetry (that
is to say, an array in which the only symmetry-equivalent reflections
are Friedel mates), we should be happy to calculate the non-anomalous
merging statistics.  At present however, we raise a Sorry because the
anomalous array is unique under symmetry, which is erroneous behaviour.