gltbx: use system libraries as fallback
When compiling with conda compilers in a conda environment the system
includes and libraries are not considered by default. This works fine
for 99% of the build, however libGL.so is not included in a conda
installation.

gltbx has a check whether libGL can be compiled against, and if it can't
then it won't be built. Here we add a fallback solution: if libGL can't
be compiled against AND we are using conda compilers AND we are on linux
then try explicitly to compile against system libraries. If this is not
successful then skip as before, otherwise build gltbx using the system
libraries.