Option to calculate add_background() on GPU
The old algorithm in C++ was dependent on the order of pixel evaluation,
therefore a new argument is provided to the C++ function
(sort_stable=true) to provide sort-order invariant results.
The CUDA version produces results very close to the C++, sort stable.
However, the match is not exact; a few dozen spots have +/- 1 photon.
Reason is unknown.