fixes memory issues with cctbx.xfel.merge
Unless you want to run an integration worker, you do
not need imagesets to  merge. In some cases,
when an imageset points to a file with many images,
it will initialize a lot of useless pointers, which
can take up lots of memory. To run an integration
worker with the merger, use the keep_imagesets=True philparam.