Remove unused file archive
As far as I can tell this was a programming jiffy from 2012 to create
minimalistic files that just about passed as HDF5 Nexus files.

In 2020 we have real data files that we can use, so this code is no
longer necessary.

This was flagged as it is the only place in cctbx_project having an
HDF5 library dependency.