No longer automatically update dependencies, #535
This pull request reverts #290 and #438. The code in these pull requests
was put in to ensure that software could declare its dependencies and
when these dependencies change (as the software is living) the
dependencies would auto-update.

People using tools where the auto-update would have done something will
now have to do this manually, however with the move to Conda this should
be rather more straightforward for the average end user.

Please see #535 if you have any questions.