For PAL NeXus, use a VDS for wavelength instead of a copy
Intial attempt was to use an ExternalLink, but this doesn't work because we need to set the units attribute and that would modify the being linked to.  Instead use a virtual dataset that we can set attributes on.

This only reduces the file size of 3000 image master from 321K to 309K but the point is to not make a copy of the wavelength array at all.  Plus, it demos how to make virtual datasets.