Adding __hash__ to sequence class
sequence (and its subclasses) objects of this module are used in other modules, such as in the function get_muscle_alignment_ordered of mmtbx/msa.py, as immutable objects and in particular as keys of dictionaries. In Python3 immutable objects are required to implement the __hash__ method (if not obvious) and the __eq__ method. So I added the __hash__ method and I can continue to use that function correctly.