Correct problem with Modulated Structures dict.
In some cases the reference obtained is not present in the dictionary. 

For example, '_geom_bond_atom_site_label_1' tag's reference is '_geom_bond_atom_site_label_' but that is not present in the Modulated Structures dictionary and it raises a KeyError: '_geom_bond_atom_site_label_'.

This change tries to correct that error.