Move nine CUDA __device__ functions to new header.
Creating a standalone header file allows the functions to be reused for independent Python extension modules that also provide CUDA code, such as in the neighboring ../gpu directory.  Functions provide for matrix/vector operations and cubic spline.

Also, the add_background CUDA kernel is removed and will be relocated in the ../gpu directory, as it is compiled into the gpu_ext.so Python extension module.