New CUDA kernel to add Bragg spots with a mask.
This affords dramatic speedup if it is enough to simulate the image on Bragg spot shoeboxes only.  For this purpose, a mask is supplied where shoebox pixels are set to True.  Key implementation points:  1) The kernel no longer loops over all pixels on the image, only over a predetermined set where the mask is True.  2) The loop over panels for multipanel detectors is no longer in C++, rather the CUDA kernel itself handles array indexing over ipanel, islow, and ifast.

Also, the add_background_CUDAKernel() is now in it’s correct place in the gpu/ directory.