debranch_maskall_CUDAKernel: move Fhkl lookup out of innermost loop
The innermost loop is over mosaic blocks. Previously, for each pixel, we
were computing the hkl indices individually for each mosaic block and
then looking up the closest structure factor. This is a lot of memory
accesses; it is unnecessary because the situation where different mosaic
blocks give different hkl indices is equivalent to saying that more than
one reflection contributes to a pixel. We are already explicitly failing
to handle overlaps, since we use only the closest Fhkl value. Therefore
we don't lose any further accuracy by looking up Fhkl only once for the
unperturbed orientation. This saves on the order of 20% of the runtime
for the kernel call.