skip unneeded memory operations in detector::each_image_allocate_cuda
Code was inefficient:  arrays were allocated and set to zero on CPU with calloc,
then allocated and copied to GPU.  Design was wasteful since these reduction arrays
are not utilized in practice.  Should they be needed in the future a faster zeroing API
must be found, and it should be encapsulated in a separate function call.

Co-authored-by: Daniel Paley <dwpaley@lbl.gov>