overload add_energy_channel_mask_allpanel_cuda to allow precomputing
active_pixel_list.  Overloaded function signature provides two choices.
flex.bool: provides a mask of active pixels (array is the same large size as the pixels)
flex.int: short list of integers are indices to whitelisted positions in the pixel array
This is the most efficient access pattern for a short list of active pixels.

Co-authored-by: Daniel Paley <dwpaley@lbl.gov>