single/multiple source in nanoBragg background.
nanoBragg background can be calculated with either monochromatic or
polychromatic (multiple source) beam.  Test cases confirm that if the
total fluence is held constant, mean intensity of the diffraction pattern
is nearly unchanged switching from mono- to narrow bandpass.
Also test that the results are identical for CPU and CUDA.

Co-authored-by: Derek Mendez <dermen@lbl.gov>