Don't divide by zero if rotating by 0°
The quaternion of a 0° rotation is, straightforwardly, 1.  Sometimes,
since the axis direction for a 0° rotation is arbitrary,
axis_and_angle_as_unit_quaternion may be called with an angle of 0 and a
placeholder axis of (0, 0, 0).  In this case, return early to guard
against a ZeroDivisionError when trying to normalise the axis.