Nasty workaround for DYLD_FALLBACK_LIBRARY_PATH
Since thanks to SIP this is clobbered in any fork / exec scenario the SHELL
default value of this is lost in the dispatcher script which is usually
harmless but in some cases - as I found - harmful.

This restores /usr/lib to the cctbx default value of that path so ensures
libraries which are depended on are found. This is not a problem in the usual
Python / conda ecosystem so this nasty workaround is probably valid and
justified.

Fixes #701