Fix simtbx flux test (#741)
Accounts for beam model change cctbx/dxtbx#488

Flux didn't used to be set in the Beam constructor, so change the cbf writing to include flux if present.  cctbx/dxtbx#493 added reading and writing of the flux in full CBFs so that will be picked up here and fix the nanoBragg test.