Fix dispatcher and console_scripts re-export on Windows
On Windows, conda uses the Scripts/ directory, instead of the bin/ directory
for dispatcher location generation. This means that on Windows we need to
use the derived script directory from get_setuptools_script_dir().

This specifically fixes the generation of dxtbx and xia dispatchers, which
relied on libtbx re-exporting anything registered under the
libtbx.dispatcher.script entry point. At some point in the past, the search
path was changed from the derived directory to a fixed bin/, which broke this
process under conda on Windows.

Fixes cctbx/dxtbx#393 and fixes xia2/xia2#594.