Add known_orientations support to the XFEL GUI
Set indexing.stills.known_orientations = the path(s) to the experiment lists containing the known crystal models, or provide the string XXX_rgYYY, where XXX is the trial number and YYY is the rungroup number.  In the latter case the XFEL GUI will replace XXX_rgYYY with the correct path.