Don't overwrite DIALS build information if already set (#753)
In cases where DIALS is resolved first in the module order, xfel previously
overwrote the env_etc entries that were already set. This change makes it so
that in the case where these variables are already configured, they won't be
overwritten.

If the .dials_dist and .dials_include aren't already written, then they will
continue to be declared by the xfel SConscript.