detector_residuals: add 2D histogram of delta psi vs. 2theta
For each reflection, compute the delta psi and two theta angles. 2D histogram is plotted where each pixel is the number of reflections with a given delta psi and two theta. Result is to a trumpet plot but for the whole dataset.

Also check for not enough reflections for the mcd filter