XFEL GUI: fix trial selection bug
Loooooongstanding bug, where if you create a new trial after visiting the run stats tab, then go back to the run stats tab and select the new trial, the new trial wouldn't update.  Only fix was restarting the GUI.  Now it updates a bit of metadata the GUI is using in the refresh_trials functio that was missing before.