Improve local_mean_intensities by renormalising the falloff with resolution. This tends to get distorted by the fact that there are different numbers of reflections at higher and lower resolution in the averaging sphere, which then causes problems using the local means for making E-values. Fixing this improves the docking performance.
Update the associated regression test to be more representative of the cryo-EM
case, i.e. make a model (new file, iotbx/regression/data/big_cube_model.pdb)
that is closer to spherical and is centered in a big cubic unit cell. Test
the performance more rigorously by checking that results are as expected from
theory.