Commit before close
The official recommendation is "cnx.commit() before cur.close()":

https://forums.mysql.com/read.php?50,578823,578941#msg-578941

but the existing behavior was to return an unclosed cursor into
a connection that might contain changes that will not be committed.
To follow the recommendation but still emulate our existing behavior, we
do these steps: (1) prefetch the results from the cursor; (2) commit or
rollback as needed; (3) close the cursor (4) return the dummy.