XFEL gui: cache database connections (#799)
As a long-standing observation, the XFEL GUI gets really slow after
a few days of data collection. One notable offender was the 'edit run
tags' dialog, which would hang for 1-2 minutes after 'OK' was pressed.

In sampling profiles collected via py-spy, the vast majority of this
hanging time was spent on repeated calls to `get_db_connection` in method
`execute_query` of `xfel.ui.db.xfel_db.db_application`. If the instance
of `db_application` saves an open database connection, then we avoid creating
a new connection for every single query.

Note this is often achieved by sharing a 'pool' of open connections
but in this case we have not noticed any problems if every `db_application`
instance just keeps its own open connection.

The existing API establishes that `db_application.execute_query`
returns a MySQLdb.Cursor into the db connection. The client was
"responsible" for closing the cursor and in general this is not done.
While it seems in general that unclosed MySQLdb cursors are harmless:

https://stackoverflow.com/questions/5669878

we maintain the encapsulation of database transactions within `db_application`
by closing the cursor and returning a `dummy_cursor` object. This prefetches results
and exposes the few methods/attributes that are currently used in client code.

Finally we enable `autocommit` on the database connection. This prevents "stale reads"
arising from unclosed transactions as discussed in:

https://stackoverflow.com/questions/1617637

Co-authored-by: Aaron Brewster <asbrewster@lbl.gov>