Implement unit tests for the whitelist mechanism.
Find strong spots on an image.  Perform nanoBragg simulation on the
strong spot shoeboxes, and a second simulation on the other pixels.
Adding the two up should equal the reference simulation on all pixels.

Test both whitelist forms: bool flags on all pixels, or whitelist pixel
address expressed in integer form.

Important change in the C++ code:  The deep array m_floatimage is now
reset to zeroes after each API call.  Otherwise the old data persists,
and would be incremented in the subsequent simulation leading to
nonsense and confusion.  Would have liked to avoid this but as we know
from information theory there is a cost to erasing information.