Change whitelist argument for addresses to size_t.
Previously, the pixel addresses for the whitelist were given in Python
by a flex.int() array.  This is now changed to flex.size_t() on the
ground that it is the more correct type for array addresses.  Address
lists can now be computed in Python with

bool_array.iselection()

rather than the more awkward

bool_array.iselection().as_int().

Corresponding changes are made in C++, CUDA and Kokkos code.