Add new plot, delta_vs_azimuthal_angle
For each reflection, plot either overall, radial, or transverse displacements vs. azimuthal angle, where the angle is between (0,1,0), i.e. the vector pointing up in real space, and the XY component of the vector from the sample to the observed spot centroid. Plot is a 2D histogram on a log scale.