modify_cosym bugfix for user-defined twin axis
Cosym internally uses a cb_op_to_minimum to put the input cell in a
standardized setting. If there is a user-defined twin axis, that also
needs to be transformed by the same cb_op. This commit fixes the
transformation for cases where the cb_op is not self-inverse (e.g. z,x,y).